+++ Sources/Sentry/SentryUIViewControllerSwizzling.m
@@ -121,17 +121,35 @@ SentryUIViewControllerSwizzling ()
 - (void)swizzleAllSubViewControllersInApp:(id<SentryUIApplication>)app
 {
     if (app.delegate == nil) {
-        NSString *message = @"UIViewControllerSwizziling: App delegate is nil. Skipping "
-                            @"swizzleAllSubViewControllersInApp.";
+        NSString *message = @"UIViewControllerSwizzling: App delegate is nil. Skipping "
+                            @"swizzling UIViewControllers in the app image.";
         [SentryLog logWithMessage:message andLevel:kSentryLevelDebug];
         return;
     }
 
-    const char *imageName = [self.objcRuntimeWrapper class_getImageName:[app.delegate class]];
+    [self swizzleUIViewControllersOfClassesInImageOf:[app.delegate class]];
+}
+
+- (void)swizzleUIViewControllersOfClassesInImageOf:(Class)class
+{
+    if (class == NULL) {
+        [SentryLog logWithMessage:@"UIViewControllerSwizzling: class is NULL. Skipping swizzling "
+                                  @"of classes in same image."
+                         andLevel:kSentryLevelDebug];
+        return;
+    }
+
+    NSString *message = [NSString
+        stringWithFormat:@"UIViewControllerSwizzling: Class to get the image name: %@", class];
+    [SentryLog logWithMessage:message andLevel:kSentryLevelDebug];
+
+    const char *imageName = [self.objcRuntimeWrapper class_getImageName:class];
 
     if (imageName == NULL) {
-        NSString *message = @"UIViewControllerSwizziling: Wasn't able to get image name of the app "
-                            @"delegate class. Skipping swizzleAllSubViewControllersInApp.";
+        NSString *message = [NSString
+            stringWithFormat:@"UIViewControllerSwizziling: Wasn't able to get image name of the "
+                             @"class: %@. Skipping swizzling of classes in same image.",
+            class];
         [SentryLog logWithMessage:message andLevel:kSentryLevelDebug];
         return;
     }
@@ -139,9 +157,16 @@ SentryUIViewControllerSwizzling ()
     NSString *appImage = [NSString stringWithCString:imageName encoding:NSUTF8StringEncoding];
 
     if (appImage == nil || appImage.length == 0) {
-        NSString *message
-            = @"UIViewControllerSwizziling: Wasn't able to get the app image name of the app "
-              @"delegate class. Skipping swizzleAllSubViewControllersInApp.";
+        NSString *message =
+            [NSString stringWithFormat:@"Wasn't able to get the app image name of the app delegate "
+                                       @"class: %@. Skipping swizzling of classes in same image.",
+                      class];
+        [SentryLog logWithMessage:message andLevel:kSentryLevelDebug];
+        return;
+    }
+
+    if ([appImage containsString:@"UIKitCore"]) {
+        NSString *message = @"UIViewControllerSwizziling: Skipping UIKitCore.";
         [SentryLog logWithMessage:message andLevel:kSentryLevelDebug];
         return;
     }
@@ -261,10 +286,15 @@ SentryUIViewControllerSwizzling ()
     for (UIViewController *viewController in allViewControllers) {
         Class viewControllerClass = [viewController class];
         if (viewControllerClass != nil) {
-            NSString *message = @"UIViewControllerSwizziling Calling swizzleRootViewController.";
-            [SentryLog logWithMessage:message andLevel:kSentryLevelDebug];
-
+            [SentryLog
+                logWithMessage:@"UIViewControllerSwizziling Calling swizzleRootViewController."
+                      andLevel:kSentryLevelDebug];
             [self swizzleViewControllerSubClass:viewControllerClass];
+
+            // We can't get the image name with the app delegate class for some apps. Therefore, we
+            // use the rootViewController and its subclasses as a fallback.  The following method
+            // ensures we don't swizzle ViewControllers of UIKit.
+            [self swizzleUIViewControllersOfClassesInImageOf:viewControllerClass];
         }
     }
 }